+++ tests/cmake.py
@@ -8,6 +8,13 @@ from pathlib import Path
 
 import pytest
 
+from .build_config import (
+    get_android_config,
+    get_platform_cmake_args,
+    get_cflags,
+    get_tsan_env,
+)
+
 
 class CMake:
     def __init__(self, factory):
@@ -122,18 +129,8 @@ def cmake(cwd, targets, options=None, cflags=None):
             "CMAKE_RUNTIME_OUTPUT_DIRECTORY_RELEASE": cwd,
         }
     )
-    if os.environ.get("ANDROID_API") and os.environ.get("ANDROID_NDK"):
-        # See: https://developer.android.com/ndk/guides/cmake
-        toolchain = "{}/ndk/{}/build/cmake/android.toolchain.cmake".format(
-            os.environ["ANDROID_HOME"], os.environ["ANDROID_NDK"]
-        )
-        options.update(
-            {
-                "CMAKE_TOOLCHAIN_FILE": toolchain,
-                "ANDROID_ABI": os.environ.get("ANDROID_ARCH") or "x86",
-                "ANDROID_NATIVE_API_LEVEL": os.environ["ANDROID_API"],
-            }
-        )
+
+    options.update(get_android_config())
 
     source_dir = os.path.dirname(os.path.dirname(os.path.realpath(__file__)))
 
@@ -154,77 +151,20 @@ def cmake(cwd, targets, options=None, cflags=None):
     config_cmd = cmake.copy()
 
     if os.environ.get("VS_GENERATOR_TOOLSET") == "ClangCL":
-        config_cmd.append("-G Visual Studio 17 2022")
-        config_cmd.append("-A x64")
-        config_cmd.append("-T ClangCL")
+        configure_clang_cl(config_cmd)
 
     for key, value in options.items():
         config_cmd.append("-D{}={}".format(key, value))
-    if sys.platform == "win32" and os.environ.get("TEST_X86"):
-        config_cmd.append("-AWin32")
-    elif sys.platform == "linux" and os.environ.get("TEST_X86"):
-        config_cmd.append("-DSENTRY_BUILD_FORCE32=ON")
-    if "asan" in os.environ.get("RUN_ANALYZER", ""):
-        config_cmd.append("-DWITH_ASAN_OPTION=ON")
-    if "tsan" in os.environ.get("RUN_ANALYZER", ""):
-        module_dir = Path(__file__).resolve().parent
-        tsan_options = {
-            "verbosity": 2,
-            "detect_deadlocks": 1,
-            "second_deadlock_stack": 1,
-            "suppressions": module_dir / "tsan.supp",
-        }
-        os.environ["TSAN_OPTIONS"] = ":".join(
-            f"{key}={value}" for key, value in tsan_options.items()
-        )
-        config_cmd.append("-DWITH_TSAN_OPTION=ON")
 
-    # we have to set `-Werror` for this cmake invocation only, otherwise
-    # completely unrelated things will break
-    if os.environ.get("ERROR_ON_WARNINGS"):
-        cflags.append("-Werror")
-    if sys.platform == "win32" and not os.environ.get("TEST_MINGW"):
-        # MP = object level parallelism, WX = warnings as errors
-        cpus = os.cpu_count()
-        cflags.append("/WX /MP{}".format(cpus))
-    if "gcc" in os.environ.get("RUN_ANALYZER", ""):
-        cflags.append("-fanalyzer")
+    config_cmd.extend(get_platform_cmake_args())
+
+    tsan_opts = get_tsan_env()
+    if tsan_opts:
+        os.environ["TSAN_OPTIONS"] = tsan_opts
+
+    cflags.extend(get_cflags())
     if "llvm-cov" in os.environ.get("RUN_ANALYZER", ""):
-        if False and os.environ.get("VS_GENERATOR_TOOLSET") == "ClangCL":
-            # for clang-cl in CI we have to use `--coverage` rather than `fprofile-instr-generate` and provide an
-            # architecture-specific profiling library for it work:
-            # TODO: This currently doesn't work due to https://gitlab.kitware.com/cmake/cmake/-/issues/24025
-            #       The issue describes a behavior where generated object-name is specified via `/fo` using a target
-            #       directory, rather than a file-name (this is CMake behavior). While the `clang-cl` suggest that this
-            #       is supported the flag produces `.gcda` and `.gcno` files, which have no relation with the object-
-            #       file and which leads to failure to accumulate coverage data.
-            #       This would have to be fixed in clang-cl: https://github.com/llvm/llvm-project/issues/87304
-            #       Let's leave this in here for posterity, it would be great to get coverage analysis on Windows.
-            flags = "--coverage"
-            profile_lib = "clang_rt.profile-x86_64.lib"
-            config_cmd.append(f"-DCMAKE_EXE_LINKER_FLAGS='{profile_lib}'")
-            config_cmd.append(f"-DCMAKE_SHARED_LINKER_FLAGS='{profile_lib}'")
-            config_cmd.append(f"-DCMAKE_MODULE_LINKER_FLAGS='{profile_lib}'")
-        else:
-            flags = "-fprofile-instr-generate -fcoverage-mapping"
-        config_cmd.append("-DCMAKE_C_FLAGS='{}'".format(flags))
-
-        # Since we overwrite `CXXFLAGS` below, we must add the experimental library here for the GHA runner that builds
-        # sentry-native with LLVM clang for macOS (to run ASAN on macOS) rather than the version coming with XCode.
-        # TODO: remove this if the GHA runner image for macOS ever updates beyond llvm15.
-        if (
-            sys.platform == "darwin"
-            and os.environ.get("CC", "") == "clang"
-            and shutil.which("clang") == "/usr/local/opt/llvm@15/bin/clang"
-        ):
-            flags = (
-                flags
-                + " -L/usr/local/opt/llvm@15/lib/c++ -fexperimental-library -Wno-unused-command-line-argument"
-            )
-
-        config_cmd.append("-DCMAKE_CXX_FLAGS='{}'".format(flags))
-    if "CMAKE_DEFINES" in os.environ:
-        config_cmd.extend(os.environ.get("CMAKE_DEFINES").split())
+        configure_llvm_cov(config_cmd)
     env = dict(os.environ)
     env["CFLAGS"] = env["CXXFLAGS"] = " ".join(cflags)
 
@@ -232,9 +172,18 @@ def cmake(cwd, targets, options=None, cflags=None):
 
     print("\n{} > {}".format(cwd, " ".join(config_cmd)), flush=True)
     try:
-        subprocess.run(config_cmd, cwd=cwd, env=env, check=True)
-    except subprocess.CalledProcessError:
-        raise pytest.fail.Exception("cmake configure failed") from None
+        result = subprocess.run(
+            config_cmd, cwd=cwd, env=env, check=True, capture_output=True, text=True
+        )
+        sys.stdout.write(result.stdout)
+        sys.stderr.write(result.stderr or "")
+    except subprocess.CalledProcessError as e:
+        if "Could NOT find ZLIB" in e.stderr:
+            pytest.skip("ZLIB not found")
+        else:
+            sys.stdout.write(e.stdout)
+            sys.stderr.write(e.stderr or "")
+            raise pytest.fail.Exception("cmake configure failed") from None
 
     # CodeChecker invocations and options are documented here:
     # https://github.com/Ericsson/codechecker/blob/master/docs/analyzer/user_guide.md
@@ -317,3 +266,45 @@ def cmake(cwd, targets, options=None, cflags=None):
             cwd=cwd,
             check=True,
         )
+
+
+def configure_clang_cl(config_cmd: list[str]):
+    config_cmd.append("-G Visual Studio 17 2022")
+    config_cmd.append("-A x64")
+    config_cmd.append("-T ClangCL")
+
+
+def configure_llvm_cov(config_cmd: list[str]):
+    if False and os.environ.get("VS_GENERATOR_TOOLSET") == "ClangCL":
+        # for clang-cl in CI we have to use `--coverage` rather than `fprofile-instr-generate` and provide an
+        # architecture-specific profiling library for it work:
+        # TODO: This currently doesn't work due to https://gitlab.kitware.com/cmake/cmake/-/issues/24025
+        #       The issue describes a behavior where generated object-name is specified via `/fo` using a target
+        #       directory, rather than a file-name (this is CMake behavior). While the `clang-cl` suggest that this
+        #       is supported the flag produces `.gcda` and `.gcno` files, which have no relation with the object-
+        #       file and which leads to failure to accumulate coverage data.
+        #       This would have to be fixed in clang-cl: https://github.com/llvm/llvm-project/issues/87304
+        #       Let's leave this in here for posterity, it would be great to get coverage analysis on Windows.
+        flags = "--coverage"
+        profile_lib = "clang_rt.profile-x86_64.lib"
+        config_cmd.append(f"-DCMAKE_EXE_LINKER_FLAGS='{profile_lib}'")
+        config_cmd.append(f"-DCMAKE_SHARED_LINKER_FLAGS='{profile_lib}'")
+        config_cmd.append(f"-DCMAKE_MODULE_LINKER_FLAGS='{profile_lib}'")
+    else:
+        flags = "-fprofile-instr-generate -fcoverage-mapping"
+    config_cmd.append("-DCMAKE_C_FLAGS='{}'".format(flags))
+
+    # Since we overwrite `CXXFLAGS` below, we must add the experimental library here for the GHA runner that builds
+    # sentry-native with LLVM clang for macOS (to run ASAN on macOS) rather than the version coming with XCode.
+    # TODO: remove this if the GHA runner image for macOS ever updates beyond llvm15.
+    if (
+        sys.platform == "darwin"
+        and os.environ.get("CC", "") == "clang"
+        and shutil.which("clang") == "/usr/local/opt/llvm@15/bin/clang"
+    ):
+        flags = (
+            flags
+            + " -L/usr/local/opt/llvm@15/lib/c++ -fexperimental-library -Wno-unused-command-line-argument"
+        )
+
+    config_cmd.append("-DCMAKE_CXX_FLAGS='{}'".format(flags))